+++ src/api/apps/easenft/dto/easenft-tx.dto.ts
@@ -1,0 +1,9 @@
+import { IsNotEmpty } from "class-validator"
+
+export class CreateTransactionDto {
+  @IsNotEmpty()
+  apiKey: string
+
+  @IsNotEmpty()
+  txId: string
+}