+++ src/api/apps/zknft/dto/update-zknft.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from "@nestjs/mapped-types"
-import { CreateZknftDto } from "./create-zknft.dto"
-
-export class UpdateZknftDto extends PartialType(CreateZknftDto) { }