+++ src/api/apps/zknft/zknft.module.ts
@@ -1,9 +1,0 @@
-import { Module } from "@nestjs/common"
-import { ZknftService } from "./zknft.service"
-import { ZknftController } from "./zknft.controller"
-
-@Module({
-  controllers: [ZknftController],
-  providers: [ZknftService],
-})
-export class ZknftModule { }